+++ _examples/chat_json/main.go
@@ -68,10 +68,11 @@ func main() {
 
 	node, _ := centrifuge.New(cfg)
 
-	engine, _ := centrifuge.NewMemoryEngine(node, centrifuge.MemoryEngineConfig{
+	// Override default broker which does not use HistoryMetaTTL.
+	engine, _ := centrifuge.NewMemoryBroker(node, centrifuge.MemoryBrokerConfig{
 		HistoryMetaTTL: 120 * time.Second,
 	})
-	node.SetEngine(engine)
+	node.SetBroker(engine)
 
 	node.OnConnecting(func(ctx context.Context, e centrifuge.ConnectEvent) (centrifuge.ConnectReply, error) {
 		cred, _ := centrifuge.GetCredentials(ctx)
